Routes added to routes folder and created public folder for express to serve static files
