+++ app.js
@@ -1,14 +1,13 @@
 const express = require('express');
 const mongoose = require('mongoose');
 const path = require('path');
-const Campground = require('./models/campground');
-const Review = require('./models/review');
 const methodOverride = require('method-override');
-const catchAsync =require('./utils/catchAsync');
 const ExpressError = require('./utils/ExpressError');
-const {campgroundSchema, reviewSchema} = require('./schema')
 const ejsMate = require('ejs-mate');
 const { request } = require('express');
+
+const campgrounds = require('./routes/campground');
+const reviews = require('./routes/reviews');
 
 mongoose.connect('mongodb://localhost:27017/yelp-camp', {
   useNewUrlParser : true,
@@ -24,89 +23,17 @@ const app =  express();
 
 app.engine('ejs', ejsMate);
 app.set('view engine', 'ejs');
-app.set('views',path.join(__dirname,'views'))
+app.set('views',path.join(__dirname,'views'));
 app.use(express.urlencoded({extended:true}));
-app.use(methodOverride('_method'))
+app.use(methodOverride('_method'));
+app.use(express.static(path.join(__dirname,'public')));
 
-const validateCampground = (req, res, next) => {
-  const { error } = campgroundSchema.validate(req.body);
-  if(error) {
-    const message = error.details.map(e => e.message).join(',');
-    throw new ExpressError(message, 400);
-  }
-  else {
-    next();
-  }
-}
-
-const validateReview = (req, res, next) =>{
-  const {error} = reviewSchema.validate(req.body);
-  if(error){
-    const message = error.details.map(e => e.message).join(',');
-    throw new ExpressError(message, 400);
-  }
-  else {
-    next();
-  }
-}
+app.use('/campgrounds', campgrounds);
+app.use('/campgrounds/:id/reviews', reviews);
 
 app.get('/', (req, res) => {
   res.render('home');
 })
-
-app.get('/campgrounds', async (req, res) => {
-  const campgrounds = await Campground.find({})
-  res.render('campgrounds/index', {campgrounds})
-})
-
-app.post('/campgrounds/', validateCampground, catchAsync(async(req,res, next) => {
-    const campground = new Campground(req.body.campground);
-    await campground.save();
-    res.redirect(`campgrounds/${campground._id}`);
-}))
-
-app.get('/campgrounds/new', (req,res) => {
-  res.render('campgrounds/new');
-})
-
-app.get('/campgrounds/:id', catchAsync(async (req, res) => {
-  const campground = await Campground.findById(req.params.id).populate('reviews')
-  res.render('campgrounds/show', {campground})
-}))
-
-app.get('/campgrounds/:id/edit', catchAsync(async (req, res) => {
-  const campground = await Campground.findById(req.params.id)
-  res.render('campgrounds/edit', {campground})
-}))
-
-app.put('/campgrounds/:id', validateCampground, catchAsync(async(req,res) =>{
-  const {id} = req.params;
-  const campground = await Campground.findByIdAndUpdate(id, {...req.body.campground})
-  res.redirect(`/campgrounds/${campground._id}`);
-}))
-
-app.delete('/campgrounds/:id', catchAsync(async(req, res) =>{
-  const {id} = req.params;
-  await Campground.findByIdAndDelete(id);
-  res.redirect('/campgrounds');
-}))
-
-app.delete('/campgrounds/:id/reviews/:reviewId', catchAsync(async(req,res)=>{
-  const {id, reviewId} = req.params;
-  const campground = await Campground.findByIdAndUpdate(id, {$pull:{reviews:reviewId}});
-  const review = await Review.findByIdAndDelete(reviewId);
-  res.redirect(`/campgrounds/${id}`);
-
-}))
-
-app.post('/campgrounds/:id/reviews', validateReview ,catchAsync(async (req, res) => {
-  const campground = await Campground.findById(req.params.id);
-  const review = new Review(req.body.review);
-  campground.reviews.push(review);
-  await review.save();
-  await campground.save();
-  res.redirect(`/campgrounds/${campground._id}`);
-}))
 
 app.all('*', (req, res, next) => {
   next(new ExpressError('Page not found', 404));
